clean map file
